+++ consul/consul.go
@@ -9,6 +9,7 @@ import (
 	"github.com/pipedrive/registrator/bridge"
 	consulapi "github.com/hashicorp/consul/api"
 	"github.com/hashicorp/go-cleanhttp"
+	"strconv"
 )
 
 const DefaultInterval = "10s"
@@ -33,16 +34,16 @@ func (f *Factory) New(uri *url.URL) bridge.RegistryAdapter {
 	if uri.Scheme == "consul-unix" {
 		config.Address = strings.TrimPrefix(uri.String(), "consul-")
 	} else if uri.Scheme == "consul-tls" {
-	        tlsConfigDesc := &consulapi.TLSConfig {
-			  Address: uri.Host,
-			  CAFile: os.Getenv("CONSUL_CACERT"),
-  			  CertFile: os.Getenv("CONSUL_TLSCERT"),
-  			  KeyFile: os.Getenv("CONSUL_TLSKEY"),
-			  InsecureSkipVerify: false,
+		tlsConfigDesc := &consulapi.TLSConfig{
+			Address:            uri.Host,
+			CAFile:             os.Getenv("CONSUL_CACERT"),
+			CertFile:           os.Getenv("CONSUL_TLSCERT"),
+			KeyFile:            os.Getenv("CONSUL_TLSKEY"),
+			InsecureSkipVerify: false,
 		}
 		tlsConfig, err := consulapi.SetupTLSConfig(tlsConfigDesc)
 		if err != nil {
-		   log.Fatal("Cannot set up Consul TLSConfig", err)
+			log.Fatal("Cannot set up Consul TLSConfig", err)
 		}
 		config.Scheme = "https"
 		transport := cleanhttp.DefaultPooledTransport()
@@ -127,6 +128,20 @@ func (r *ConsulAdapter) buildCheck(service *bridge.Service) *consulapi.AgentServ
 
 func (r *ConsulAdapter) Deregister(service *bridge.Service) error {
 	return r.client.Agent().ServiceDeregister(service.ID)
+}
+
+func (r *ConsulAdapter) SetupHealthCheck(service *bridge.Service, healthCheck *bridge.TtlHealthCheck) error {
+	ttlCheck := new(consulapi.AgentCheckRegistration)
+	ttlCheck.ID = service.ID + "-gracefull-shutdown"
+	ttlCheck.Name = "Gracefully shutting down " + service.ID
+	ttlCheck.Notes = "TTL check to wait until service " + service.ID + " will gracefully shutdown"
+	ttlCheck.ServiceID = service.ID
+	ttlCheck.AgentServiceCheck = consulapi.AgentServiceCheck{TTL: strconv.Itoa(healthCheck.TTL) + "s"}
+	err := r.client.Agent().CheckRegister(ttlCheck)
+	if err != nil {
+		return err
+	}
+	return r.client.Agent().UpdateTTL(ttlCheck.ID, "Graceful shutdown in progress", healthCheck.CheckStatus)
 }
 
 func (r *ConsulAdapter) Refresh(service *bridge.Service) error {